+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "dotenv": "^14.2.0",
     "express": "^4.17.2",
     "mongoose": "^6.1.7"
   },
